+++ exporters/stdout/stdoutmetric/metric_test.go
@@ -23,12 +23,11 @@ import (
 	"testing"
 	"time"
 
-	"go.opentelemetry.io/otel/exporters/stdout/stdoutmetric"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/stdout/stdoutmetric"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
 	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"